refactor(#131 WI-7b): extract EpubBilingualInjectJs to keep EpubBilingualJs under ~300 lines
Behavior-preserving split: the inject+reconcile builder (the largest branching script)
moves to EpubBilingualInjectJs.kt as an internal EpubBilingualJs extension. EpubBilingualJs.kt
242 lines, EpubBilingualInjectJs.kt 122. No behavior change; JVM tests still green.

Refs #131
